Backend/requirements.txt
pandas
numpy
scikit-learn
flask
flask-cors
pymongo
bcrypt
python-dotenv
